Add files via upload
This program analyzes the US-EPA air quality or meteo scalar files in
combination with Wind speed and direction files downloaded using the get_EPA_files
script. The result of analysis is the list of stations and instruments with
mutual Wind and specified Variable data (e.g. PM10) available for each year
in a defined period. This is useful for determining the period in which the
time series of wind and scalar data is available at the station of interest.
The list is stored in the output file (e.g. Analyze_EPA_Files.csv).


The user defines the: 1) start and end year
                  	  2) path to directory with files
                  	  3) variable prefixes in file names (i.e.for PM10_1991.csv,
                                                  	prefix is 'PM10')
              	      4) output file name
    
The program finds and makes a list of all stations and instruments
(identified with POC number) and stores the list in output csv file.

Usage:
Download US EPA files using get_EPA_files and edit SET OPTIONS: